Refactored engine usage

+++ dds_glossary/database.py
@@ -1,5 +1,8 @@
 """Database classes for the dds_glossary package."""
 
+from os import getenv
+from typing import Optional
+
 from sqlalchemy import ForeignKey, create_engine
 from sqlalchemy.engine import Engine
 from sqlalchemy.orm import (
@@ -79,25 +82,37 @@ class Relationship(Base):  # pylint: disable=too-few-public-methods
     )
 
 
-def init_engine(database_url: str, drop_database_flag: bool = False) -> Engine:
+def init_engine(
+    database_url: Optional[str] = None,
+    drop_database_flag: bool = False,
+) -> Engine:
     """
-    Initialize the database engine for the given `database_url`. If the
-    database does not exist, it will be created. If it does exist, it will be
-    dropped and recreated if the `drop_database_flag` argument is set to
-    `True`. The database schema will also be created.
+    Initialize the database engine. If the database does not exist, create it.
+    If the drop_database_flag is set, drop the database if it exists. Create the
+    tables if they do not exist.
 
     Args:
-        database_url (str): The database URL.
-        drop_database_flag (bool): Whether to drop the database before
-            creating it. Defaults to `False`.
+        database_url (Optional[str]): The database URL. If None, use the
+            DATABASE_URL environment variable.
+        drop_database_flag (bool): Flag to drop the database if it exists.
 
     Returns:
         Engine: The database engine.
+
+    Raises:
+        ValueError: If the DATABASE_URL environment variable is not set.
     """
+    if database_url is None:
+        database_url = getenv("DATABASE_URL")
+
+    if not database_url:
+        raise ValueError("DATABASE_URL environment variable not set.")
+
     if database_exists(database_url) and drop_database_flag:
         drop_database(database_url)
     if not database_exists(database_url):
         create_database(database_url)
+
     engine = create_engine(database_url)
     Base.metadata.create_all(engine)
     return engine
